+++ app/src/main/java/com/texthip/thip/ui/navigator/extensions/FeedNavigationExtensions.kt
@@ -16,12 +16,13 @@ fun NavHostController.navigateToMySubscription() {
 
 // 피드 작성으로
 fun NavHostController.navigateToFeedWrite() {
-    navigate(FeedRoutes.Write)
+    navigate(FeedRoutes.Write())
 }
 
 // 피드 댓글으로
 fun NavHostController.navigateToFeedComment(feedId: Int) {
     navigate(FeedRoutes.Comment(feedId))
+}
 
 fun NavHostController.navigateToFeedWrite(
     isbn: String? = null,
